# Improve performance with Preallocated buffer and hardcoded metering opcodes