--- a/package.json
+++ b/package.json
@@ -3,7 +3,7 @@
 	"type": "module",
 	"module": "src/index.js",
 	"main": "src/index.js",
-	"version": "2.0.0",
+	"version": "2.0.1",
 	"scripts": {
 		"fmt": "bunx biome check --write ."
 	},