--- a/src/wasm2json.js
+++ b/src/wasm2json.js
@@ -1,10 +1,10 @@
-import Stream from "buffer-pipe";
 /**
  * This code is an adaptation over https://github.com/warp-contracts/warp-wasm-json-toolkit/blob/main/wasm2json.js.
  *
  * License: MPL-2.0
  */
 import leb from "leb128";
+import { Stream } from "./buffer-pipe.js";
 import { OP_IMMEDIATES } from "./immediates.js";
 
 const wasm2json = (buf, filter) => {