## Motivation

Speeds up the compilation and injecting the metering code significantly.

## Explanation of Changes

Went for the easy fixes first: use a buffer that aggressively preallocates and grows as needed, and remove text parsing code.

I think we could do a lot more, but these simple changes are already a huge improvement.

## Testing

Added a unit test to keep the speed in check, as well as running benchmark tests on the `test-vm.wasm`

```js
import fs from "node:fs";
import { resolve } from "node:path";
import { costTable } from "./src/cost-table-example.js";
import { meterWasm } from "./src/index.js";

const input = fs.readFileSync(
	resolve("../seda-sdk/libs/wasm-integration-tests/test-vm.wasm"),
);

const iterations = 50;
let totalTime = 0;

for (let i = 0; i < iterations; i++) {
	const start = process.hrtime.bigint();
	const result = meterWasm(input, costTable);
	new WebAssembly.Module(result);
	const end = process.hrtime.bigint();
	totalTime += Number(end - start);
}

const avgTimeMs = totalTime / iterations / 1e6;
console.log(
	`Average time per run: ${avgTimeMs.toFixed(3)} ms over ${iterations} iterations`,
);
```

Before: `Average time per run: 503.991 ms over 50 iterations`
After: `Average time per run: 127.012 ms over 50 iterations`

`console.time` for the big sections:
```
[66.84ms] wasm2json
[0.07ms] injectMemoryLimit
[53.31ms] meterJSON
[77.45ms] json2wasm
```

They speed up on subsequent runs but not sure if that's relevant in our case as well.

## Related PRs and Issues

N.A.